requirements.txt
appdirs==1.4.3
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
Django==3.0.6
django-cors-headers==3.2.1
djangorestframework==3.11.0
idna==2.9
imagecodecs==2020.2.18
imageio==2.8.0
imutils==0.5.3
joblib==0.14.1
kiwisolver==1.2.0
matplotlib==3.2.1
mysqlclient==1.4.6
networkx==2.4
numpy==1.18.4
opencv-python==4.2.0.34
packaging==20.3
Pillow==7.1.2
pooch==1.1.0
pyparsing==2.4.7
pytesseract==0.3.4
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
requests==2.23.0
scikit-image==0.17.1
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
tifffile==2020.5.7
urllib3==1.25.9
